通用提交

+++ src/main/java/com/jiagouedu/mapper/UserMapper.java
@@ -3,13 +3,13 @@ package com.jiagouedu.mapper;
 import com.jiagouedu.pojo.User;
 
 public interface UserMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(User record);
 
     int insertSelective(User record);
 
-    User selectByPrimaryKey(Integer id);
+    User selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(User record);
